+++ src/intTest/com/nftco/flow/sdk/IntegrationTestUtils.kt
@@ -1,0 +1,24 @@
+package com.nftco.flow.sdk
+
+object IntegrationTestUtils {
+    fun newMainnetAccessApi(): FlowAccessApi = Flow.newAccessApi(MAINNET_HOSTNAME)
+
+    fun newTestnetAccessApi(): FlowAccessApi = Flow.newAccessApi(TESTNET_HOSTNAME)
+
+    private const val MAINNET_HOSTNAME = "access.mainnet.nodes.onflow.org"
+    private const val TESTNET_HOSTNAME = "access.devnet.nodes.onflow.org"
+
+    var transaction = FlowTransaction(
+        script = FlowScript("import 0xsomething \n {}"),
+        arguments = listOf(FlowArgument(byteArrayOf(2, 2, 3)), FlowArgument(byteArrayOf(3, 3, 3))),
+        referenceBlockId = FlowId.of(byteArrayOf(3, 3, 3, 6, 6, 6)),
+        gasLimit = 44,
+        proposalKey = FlowTransactionProposalKey(
+            address = FlowAddress.of(byteArrayOf(4, 5, 4, 5, 4, 5)),
+            keyIndex = 11,
+            sequenceNumber = 7
+        ),
+        payerAddress = FlowAddress.of(byteArrayOf(6, 5, 4, 3, 2)),
+        authorizers = listOf(FlowAddress.of(byteArrayOf(9, 9, 9, 9, 9)), FlowAddress.of(byteArrayOf(8, 9, 9, 9, 9)))
+    )
+}
